test(data-model): mirror src/ subdir structure + split tests by source
Reorganize `packages/data-model/test/` to parallel the `src/` subdir
structure (`test/fabric-instances/`, `test/fabric-primitives/`,
`test/json-wire/`) and split multi-source test files into per-source
files matching the 080 by-class + #3678 json-wire decompositions.

- Clean moves: FabricBytes/FabricHash -> test/fabric-primitives/;
  fabric-regexp.test.ts -> test/fabric-instances/FabricRegExp.test.ts
  (export-named); json-encoding/JsonEncodingContext -> test/json-wire/.
- fabric-epoch.test.ts split -> FabricEpochNsec.test.ts +
  FabricEpochDays.test.ts (per the split src) + a small fabric-epoch.test.ts
  cross-check for the both-classes integration test.
- fabric-native-instances.test.ts split per-class ->
  FabricError/FabricMap/FabricSet/UnknownValue/ProblematicValue/
  ExplicitTagValue.test.ts, each absorbing its `[DEEP_FREEZE]` matrix
  blocks under one top-level `describe`; the FabricRegExp matrix blocks
  fold into FabricRegExp.test.ts; the non-class function tests
  (native-instance-utils + cross-cutting matrices) -> a residual
  native-instance-utils.test.ts.
- Shared fixtures (DummyReconstructionContext, dummyContext, the
  protocol recursion-callback helpers) -> test/fabric-instances/fixtures.ts
  (content-hash fixtures.ts precedent).
- deno.json test glob `test/*.test.ts` -> `test/**/*.test.ts` so the
  subdir tests run.

Conservation of content: every `it()` body moves verbatim (modulo the
de-nest re-indentation forced by the single-top-level-`describe()`
restructure); only import paths repoint and the top-level `describe()`
wrapping is retitled per source. Test count conserved (1123 `it()`s
before and after); suite 50 passed (1714 steps) / 0 failed.
