+++ packages/data-model/deno.json
@@ -20,7 +20,7 @@
   },
   "test": {
     "include": [
-      "test/*.test.ts"
+      "test/**/*.test.ts"
     ]
   },
   "exports": {
