+++ packages/data-model/test/deep-freeze.test.ts
@@ -336,7 +336,7 @@ describe("isDeepFrozenFabricValue with FabricInstance (R6)", () => {
 // Arm 2 (`FabricPrimitive`) are structurally cycle-free (leaf / no outbound
 // references), so cycle tests only apply to Arm 4 (plain-object / array
 // fallback) here. Arm 3 (`FabricInstance` via `[DEEP_FREEZE]`) cycles are
-// covered in `fabric-native-instances.test.ts`.
+// covered in `fabric-instances/native-instance-utils.test.ts`.
 //
 // Termination assertion: a cycle without shared-`inProgress` threading
 // would manifest as `RangeError: Maximum call stack size exceeded` (a
