+++ packages/data-model/test/fabric-primitives/FabricEpochNsec.test.ts
@@ -1,0 +1,49 @@
+import { describe, it } from "@std/testing/bdd";
+import { expect } from "@std/expect";
+import { FabricEpochNsec } from "../../src/fabric-primitives/FabricEpochNsec.ts";
+import { FabricInstance, FabricPrimitive } from "../../src/interface.ts";
+
+describe("FabricEpochNsec", () => {
+  it("wraps a bigint value", () => {
+    const sn = new FabricEpochNsec(1234567890000000000n);
+    expect(sn.value).toBe(1234567890000000000n);
+  });
+
+  it("wraps zero", () => {
+    const sn = new FabricEpochNsec(0n);
+    expect(sn.value).toBe(0n);
+  });
+
+  it("wraps negative values (pre-epoch)", () => {
+    const sn = new FabricEpochNsec(-1000000000n);
+    expect(sn.value).toBe(-1000000000n);
+  });
+
+  it("handles large future date (year 3000)", () => {
+    const nsec = 32503680000000000000n;
+    const sn = new FabricEpochNsec(nsec);
+    expect(sn.value).toBe(nsec);
+  });
+
+  it("is instanceof FabricEpochNsec", () => {
+    const sn = new FabricEpochNsec(42n);
+    expect(sn instanceof FabricEpochNsec).toBe(true);
+  });
+
+  it("is instanceof FabricPrimitive", () => {
+    expect(new FabricEpochNsec(0n) instanceof FabricPrimitive).toBe(
+      true,
+    );
+  });
+
+  it("instances are always frozen", () => {
+    expect(Object.isFrozen(new FabricEpochNsec(42n))).toBe(true);
+  });
+
+  describe("protocol", () => {
+    it("is NOT a FabricInstance (no DECONSTRUCT)", () => {
+      const sn = new FabricEpochNsec(0n);
+      expect(sn instanceof FabricInstance).toBe(false);
+    });
+  });
+});
