+++ packages/data-model/test/fabric-primitives/FabricHash.test.ts
@@ -1,6 +1,6 @@
 import { describe, it } from "@std/testing/bdd";
 import { expect } from "@std/expect";
-import { FabricHash } from "../src/fabric-primitives/FabricHash.ts";
+import { FabricHash } from "../../src/fabric-primitives/FabricHash.ts";
 
 /** A fixed 32-byte hash for deterministic tests. */
 const SAMPLE_HASH = new Uint8Array(32);
