+++ packages/data-model/test/json-wire/JsonEncodingContext.test.ts
@@ -1,20 +1,20 @@
 import { describe, it } from "@std/testing/bdd";
 import { expect } from "@std/expect";
-import { JsonEncodingContext } from "../src/json-wire/JsonEncodingContext.ts";
-import type { FabricValue } from "../src/interface.ts";
-import type { JsonWireValue } from "../src/json-wire/interface.ts";
-import { UnknownValue } from "../src/fabric-instances/UnknownValue.ts";
-import { ProblematicValue } from "../src/fabric-instances/ProblematicValue.ts";
-import { FabricEpochDays } from "../src/fabric-primitives/FabricEpochDays.ts";
-import { FabricEpochNsec } from "../src/fabric-primitives/FabricEpochNsec.ts";
-import { FabricError } from "../src/fabric-instances/FabricError.ts";
-import { isDeepFrozen } from "../src/deep-freeze.ts";
-import { BaseReconstructionContext } from "../src/BaseReconstructionContext.ts";
+import { JsonEncodingContext } from "../../src/json-wire/JsonEncodingContext.ts";
+import type { FabricValue } from "../../src/interface.ts";
+import type { JsonWireValue } from "../../src/json-wire/interface.ts";
+import { UnknownValue } from "../../src/fabric-instances/UnknownValue.ts";
+import { ProblematicValue } from "../../src/fabric-instances/ProblematicValue.ts";
+import { FabricEpochDays } from "../../src/fabric-primitives/FabricEpochDays.ts";
+import { FabricEpochNsec } from "../../src/fabric-primitives/FabricEpochNsec.ts";
+import { FabricError } from "../../src/fabric-instances/FabricError.ts";
+import { isDeepFrozen } from "../../src/deep-freeze.ts";
+import { BaseReconstructionContext } from "../../src/BaseReconstructionContext.ts";
 import {
   resetDataModelConfig,
   setDataModelConfig,
   shallowFabricFromNativeValue,
-} from "../src/fabric-value.ts";
+} from "../../src/fabric-value.ts";
 
 /**
  * Shared test `ReconstructionContext`: `getCell()` always throws (no test
